Conflicts:
	TeachEasy/src/teacheasy/render/Renderer.java

+++ TeachEasy/src/teacheasy/render/Renderer.java
@@ -168,12 +168,12 @@ public class Renderer {
     
     /** Render a video on a page */
     private void renderVideo(VideoObject video) {
-        videoHandler.createVideo(bounds.getMaxX() * video.getXStart(),
-                                 bounds.getMaxY() * video.getYStart(),
-                                 800,
-                                 video.getSourcefile(),
-                                 false,
-                                 false);
+    	videoHandler.createVideo((float)bounds.getMaxX() * video.getXStart(),
+    							(float)bounds.getMaxY() * video.getYStart(),
+							    800,
+							    video.getSourcefile(),
+							    false,
+							    false);
     }
     
     /** Render an audio object on a page */
